# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Diruma",
        "description": "",
        "contact": {
            "email": "email@anda.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/contact/update-info": {
            "post": {
                "tags": [
                    "Contact"
                ],
                "summary": "Update Info Dasar (Alamat, Telp, Jam Buka) - Admin Only",
                "operationId": "cea9994565119f1be2f69a04feb390e3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "address": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "working_hour": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil Update"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/contact/social-media": {
            "post": {
                "tags": [
                    "Contact"
                ],
                "summary": "Tambah Sosial Media Baru (Admin Only)",
                "operationId": "a280e05e1f89bbd02b4a9d5433dc0a25",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "handle": {
                                        "type": "string",
                                        "example": " "
                                    },
                                    "social_media": {
                                        "type": "string",
                                        "example": " "
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil Ditambah"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/contact/social-media/{id}": {
            "delete": {
                "tags": [
                    "Contact"
                ],
                "summary": "Hapus Sosial Media (Admin Only)",
                "operationId": "28f0fb5f3c385927267abbde7e00e3d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil Dihapus"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/menus": {
            "get": {
                "tags": [
                    "Menu"
                ],
                "summary": "Ambil Semua Data Menu",
                "description": "Endpoint ini membutuhkan Token JWT",
                "operationId": "54c2d937e1a32b13f880b49ce0444f34",
                "responses": {
                    "200": {
                        "description": "Sukses ambil data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Sukses"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token Salah / Tidak Ada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/menus/{id}": {
            "get": {
                "tags": [
                    "Menu"
                ],
                "summary": "Lihat Detail Satu Menu",
                "operationId": "f21bb63f9fe1ea437a502e5116af01a8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Menu",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail Ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "string",
                                            "example": "Sukses"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Menu Tidak Ditemukan"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login dengan email dan password untuk dapatkan token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}